+++ addOns/automation/src/test/java/org/zaproxy/addon/automation/jobs/PolicyDefinitionUnitTest.java
@@ -368,4 +368,107 @@ class PolicyDefinitionUnitTest {
         // Then
         assertThat(ruleYaml, containsString("id: " + id));
     }
+
+    @ParameterizedTest
+    @ValueSource(strings = {"TEST_TAG", "TEST_.*"})
+    void shouldAddRuleUsingAlertTags(String tagPattern) {
+        // Given
+        String yamlStr =
+                String.format(
+                        """
+                defaultStrength: low
+                defaultThreshold: 'off'
+                alertTags:
+                - include:
+                    - %s
+                  exclude: []
+                  strength: insane
+                  threshold: high
+                """,
+                        tagPattern);
+        AutomationProgress progress = new AutomationProgress();
+        Yaml yaml = new Yaml();
+        Object data = yaml.load(yamlStr);
+
+        // When
+        policyDefinition.parsePolicyDefinition(data, "test", progress);
+
+        // Then
+        assertThat(progress.hasErrors(), is(equalTo(false)));
+        assertThat(progress.hasWarnings(), is(equalTo(false)));
+        assertThat(policyDefinition.getDefaultStrength(), is(equalTo("low")));
+        assertThat(policyDefinition.getDefaultThreshold(), is(equalTo("off")));
+        List<Rule> rules = policyDefinition.getEffectiveRules();
+        assertThat(rules.size(), is(equalTo(1)));
+        assertThat(rules.get(0).getId(), is(equalTo(50000)));
+        assertThat(rules.get(0).getName(), is(equalTo("PluginTestHelper")));
+        assertThat(rules.get(0).getStrength(), is(equalTo("insane")));
+        assertThat(rules.get(0).getThreshold(), is(equalTo("high")));
+    }
+
+    @ParameterizedTest
+    @ValueSource(strings = {"TEST_TAG", "TEST_.*"})
+    void shouldExcludeIncludedRulesUsingAlertTags(String tagPattern) {
+        // Given
+        String yamlStr =
+                String.format(
+                        """
+                defaultStrength: low
+                defaultThreshold: medium
+                alertTags:
+                - include:
+                  - .*
+                  exclude:
+                  - %s
+                """,
+                        tagPattern);
+        AutomationProgress progress = new AutomationProgress();
+        Yaml yaml = new Yaml();
+        Object data = yaml.load(yamlStr);
+
+        // When
+        policyDefinition.parsePolicyDefinition(data, "test", progress);
+
+        // Then
+        assertThat(progress.hasErrors(), is(equalTo(false)));
+        assertThat(progress.hasWarnings(), is(equalTo(false)));
+        assertThat(policyDefinition.getEffectiveRules().isEmpty(), is(equalTo(true)));
+    }
+
+    @Test
+    void shouldNotAddSameRuleTwice() {
+        // Given
+        String yamlStr =
+                """
+                defaultStrength: low
+                defaultThreshold: 'off'
+                rules:
+                - id: 50000
+                  name: rule1
+                  strength: insane
+                  threshold: high
+                alertTags:
+                - include:
+                    - TEST_TAG
+                  exclude: []
+                  strength: low
+                  threshold: medium
+                """;
+        AutomationProgress progress = new AutomationProgress();
+        Yaml yaml = new Yaml();
+        Object data = yaml.load(yamlStr);
+
+        // When
+        policyDefinition.parsePolicyDefinition(data, "test", progress);
+
+        // Then
+        assertThat(progress.hasErrors(), is(equalTo(false)));
+        assertThat(progress.hasWarnings(), is(equalTo(false)));
+        List<Rule> rules = policyDefinition.getEffectiveRules();
+        assertThat(rules.size(), is(equalTo(1)));
+        assertThat(rules.get(0).getId(), is(equalTo(50000)));
+        assertThat(rules.get(0).getName(), is(equalTo("PluginTestHelper")));
+        assertThat(rules.get(0).getStrength(), is(equalTo("insane")));
+        assertThat(rules.get(0).getThreshold(), is(equalTo("high")));
+    }
 }